Do not extract reactions if product energy is higher than previous node
If the product energy is larger than the energy of the previous node,
then something weird probably happened during GSM, so discard the file.
However, if the product energy is very large, then there probably just
was an error during the product optimization, but the structure is
probably good enough for SMILES perception.